requirements.txt
discord.py
python-dotenv
emoji
disputils